requirements.txt
Django>=3.0,<3.2.10
psycopg2-binary>=2.9.2
django-mptt
django-elasticsearch-dsl